mark as favorite

+++ app/src/main/java/com/moviesearcher/adapters/FavoriteMoviesAdapter.kt
@@ -52,7 +52,7 @@ class FavoriteMoviesAdapter(
 
             navController.navigate(
                 FavoritesFragmentDirections.actionFragmentFavoritesToMovieInfoFragment(
-                    movieId
+                    movieId.toLong()
                 )
             )
         }
